requirements.txt
requests
moviepy
google-cloud-texttospeech
google-cloud-speech
pydub
python-dotenv
ffmpeg-python

